♻️(backend) split video seriaializers files
As video serializer file grows, we will soon reach file length limit.
All of them has been splitted on their own dedicated file.
